Lots of bug fixes
There's now checks to see if a server or member from the birthday list exist.
remove_user_bd() wrote on the channel date file instead of user.
On first boot, it won't go through the birthday list to send messages.
get_user_info() was a coroutine, I added an await.
SetUserBd no longer allows the creation of infinite amounts of UserBirthday.

ListUserBD now displays the dates in order of closest to today.